pyproject.toml
[tool.poetry]
name = "bazelbot"
version = "0.1.0"
description = ""
authors = ["Robin De Smet <robin_de_smet@hotmail.be>"]
readme = "README.md"
packages = [{ include = "src", from = "./" }]

[tool.poetry.dependencies]
python = "^3.12"
llama-index = "^0.10.27"
ollama = "^0.1.8"
python-dotenv = "^1.0.1"
pydantic = "^2.7.1"
discord-py = "^2.3.2"
llama-index-llms-ollama = "^0.1.3"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
pylint = "^3.2.0"
scikit-learn = "^1.5.0"
google-generativeai = "^0.8.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = [
    "llm: Tests where we need the LLM.",
]